requirements.txt
langchain
faiss-cpu
sentence-transformers
mistralai
python-dotenv
PyPDF2
streamlit
